+++ src/aind_dynamic_foraging_models/generative_model/params/forager_loss_counting_params.py
@@ -5,7 +5,7 @@ from typing import Literal, Tuple, Type
 
 from pydantic import BaseModel, Field
 
-from .base import create_pydantic_models_dynamic
+from .util import create_pydantic_models_dynamic
 from .forager_q_learning_params import _add_choice_kernel_fields
